+++ src/Racmacs_wrappers_sera.cpp
@@ -25,6 +25,8 @@ std::string ac_sr_get_passage( const AcSerum &sr ){ return sr.get_passage(); }
 // [[Rcpp::export(rng = false)]]
 std::vector<std::string> ac_sr_get_clade( const AcSerum &sr ){ return sr.get_clade(); }
 // [[Rcpp::export(rng = false)]]
+std::vector<std::string> ac_sr_get_annotations( const AcSerum &sr ){ return sr.get_annotations(); }
+// [[Rcpp::export(rng = false)]]
 std::string ac_sr_get_name( const AcSerum &sr ){ return sr.get_name(); }
 // [[Rcpp::export(rng = false)]]
 std::string ac_sr_get_name_full( const AcSerum &sr ){ return sr.get_name_full(); }
@@ -53,6 +55,8 @@ AcSerum ac_sr_set_reference( AcSerum sr, bool value ){  sr.set_reference(value);
 AcSerum ac_sr_set_passage( AcSerum sr, std::string value ){ sr.set_passage(value); return sr; }
 // [[Rcpp::export(rng = false)]]
 AcSerum ac_sr_set_clade( AcSerum sr, std::vector<std::string> value ){ sr.set_clade(value); return sr; }
+// [[Rcpp::export(rng = false)]]
+AcSerum ac_sr_set_annotations( AcSerum sr, std::vector<std::string> value ){ sr.set_annotations(value); return sr; }
 // [[Rcpp::export(rng = false)]]
 AcSerum ac_sr_set_name( AcSerum sr, std::string value ){ sr.set_name(value); return sr; }
 // [[Rcpp::export(rng = false)]]