+++ src/RcppExports.cpp
@@ -256,6 +256,26 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// ac_ag_get_annotations
+std::vector<std::string> ac_ag_get_annotations(const AcAntigen& ag);
+RcppExport SEXP _Racmacs_ac_ag_get_annotations(SEXP agSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< const AcAntigen& >::type ag(agSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_ag_get_annotations(ag));
+    return rcpp_result_gen;
+END_RCPP
+}
+// ac_ag_get_labids
+std::vector<std::string> ac_ag_get_labids(const AcAntigen& ag);
+RcppExport SEXP _Racmacs_ac_ag_get_labids(SEXP agSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< const AcAntigen& >::type ag(agSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_ag_get_labids(ag));
+    return rcpp_result_gen;
+END_RCPP
+}
 // ac_ag_get_name
 std::string ac_ag_get_name(const AcAntigen& ag);
 RcppExport SEXP _Racmacs_ac_ag_get_name(SEXP agSEXP) {
@@ -378,6 +398,28 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< AcAntigen >::type ag(agSEXP);
     Rcpp::traits::input_parameter< std::vector<std::string> >::type value(valueSEXP);
     rcpp_result_gen = Rcpp::wrap(ac_ag_set_clade(ag, value));
+    return rcpp_result_gen;
+END_RCPP
+}
+// ac_ag_set_annotations
+AcAntigen ac_ag_set_annotations(AcAntigen ag, std::vector<std::string> value);
+RcppExport SEXP _Racmacs_ac_ag_set_annotations(SEXP agSEXP, SEXP valueSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< AcAntigen >::type ag(agSEXP);
+    Rcpp::traits::input_parameter< std::vector<std::string> >::type value(valueSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_ag_set_annotations(ag, value));
+    return rcpp_result_gen;
+END_RCPP
+}
+// ac_ag_set_labids
+AcAntigen ac_ag_set_labids(AcAntigen ag, std::vector<std::string> value);
+RcppExport SEXP _Racmacs_ac_ag_set_labids(SEXP agSEXP, SEXP valueSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< AcAntigen >::type ag(agSEXP);
+    Rcpp::traits::input_parameter< std::vector<std::string> >::type value(valueSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_ag_set_labids(ag, value));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -919,6 +961,16 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// ac_sr_get_annotations
+std::vector<std::string> ac_sr_get_annotations(const AcSerum& sr);
+RcppExport SEXP _Racmacs_ac_sr_get_annotations(SEXP srSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< const AcSerum& >::type sr(srSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_sr_get_annotations(sr));
+    return rcpp_result_gen;
+END_RCPP
+}
 // ac_sr_get_name
 std::string ac_sr_get_name(const AcSerum& sr);
 RcppExport SEXP _Racmacs_ac_sr_get_name(SEXP srSEXP) {
@@ -1041,6 +1093,17 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< AcSerum >::type sr(srSEXP);
     Rcpp::traits::input_parameter< std::vector<std::string> >::type value(valueSEXP);
     rcpp_result_gen = Rcpp::wrap(ac_sr_set_clade(sr, value));
+    return rcpp_result_gen;
+END_RCPP
+}
+// ac_sr_set_annotations
+AcSerum ac_sr_set_annotations(AcSerum sr, std::vector<std::string> value);
+RcppExport SEXP _Racmacs_ac_sr_set_annotations(SEXP srSEXP, SEXP valueSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< AcSerum >::type sr(srSEXP);
+    Rcpp::traits::input_parameter< std::vector<std::string> >::type value(valueSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_sr_set_annotations(sr, value));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1650,6 +1713,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_ag_get_reference", (DL_FUNC) &_Racmacs_ac_ag_get_reference, 1},
     {"_Racmacs_ac_ag_get_passage", (DL_FUNC) &_Racmacs_ac_ag_get_passage, 1},
     {"_Racmacs_ac_ag_get_clade", (DL_FUNC) &_Racmacs_ac_ag_get_clade, 1},
+    {"_Racmacs_ac_ag_get_annotations", (DL_FUNC) &_Racmacs_ac_ag_get_annotations, 1},
+    {"_Racmacs_ac_ag_get_labids", (DL_FUNC) &_Racmacs_ac_ag_get_labids, 1},
     {"_Racmacs_ac_ag_get_name", (DL_FUNC) &_Racmacs_ac_ag_get_name, 1},
     {"_Racmacs_ac_ag_get_name_full", (DL_FUNC) &_Racmacs_ac_ag_get_name_full, 1},
     {"_Racmacs_ac_ag_get_name_abbreviated", (DL_FUNC) &_Racmacs_ac_ag_get_name_abbreviated, 1},
@@ -1662,6 +1727,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_ag_set_reference", (DL_FUNC) &_Racmacs_ac_ag_set_reference, 2},
     {"_Racmacs_ac_ag_set_passage", (DL_FUNC) &_Racmacs_ac_ag_set_passage, 2},
     {"_Racmacs_ac_ag_set_clade", (DL_FUNC) &_Racmacs_ac_ag_set_clade, 2},
+    {"_Racmacs_ac_ag_set_annotations", (DL_FUNC) &_Racmacs_ac_ag_set_annotations, 2},
+    {"_Racmacs_ac_ag_set_labids", (DL_FUNC) &_Racmacs_ac_ag_set_labids, 2},
     {"_Racmacs_ac_ag_set_name", (DL_FUNC) &_Racmacs_ac_ag_set_name, 2},
     {"_Racmacs_ac_ag_set_name_full", (DL_FUNC) &_Racmacs_ac_ag_set_name_full, 2},
     {"_Racmacs_ac_ag_set_name_abbreviated", (DL_FUNC) &_Racmacs_ac_ag_set_name_abbreviated, 2},
@@ -1713,6 +1780,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_sr_get_reference", (DL_FUNC) &_Racmacs_ac_sr_get_reference, 1},
     {"_Racmacs_ac_sr_get_passage", (DL_FUNC) &_Racmacs_ac_sr_get_passage, 1},
     {"_Racmacs_ac_sr_get_clade", (DL_FUNC) &_Racmacs_ac_sr_get_clade, 1},
+    {"_Racmacs_ac_sr_get_annotations", (DL_FUNC) &_Racmacs_ac_sr_get_annotations, 1},
     {"_Racmacs_ac_sr_get_name", (DL_FUNC) &_Racmacs_ac_sr_get_name, 1},
     {"_Racmacs_ac_sr_get_name_full", (DL_FUNC) &_Racmacs_ac_sr_get_name_full, 1},
     {"_Racmacs_ac_sr_get_name_abbreviated", (DL_FUNC) &_Racmacs_ac_sr_get_name_abbreviated, 1},
@@ -1725,6 +1793,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_sr_set_reference", (DL_FUNC) &_Racmacs_ac_sr_set_reference, 2},
     {"_Racmacs_ac_sr_set_passage", (DL_FUNC) &_Racmacs_ac_sr_set_passage, 2},
     {"_Racmacs_ac_sr_set_clade", (DL_FUNC) &_Racmacs_ac_sr_set_clade, 2},
+    {"_Racmacs_ac_sr_set_annotations", (DL_FUNC) &_Racmacs_ac_sr_set_annotations, 2},
     {"_Racmacs_ac_sr_set_name", (DL_FUNC) &_Racmacs_ac_sr_set_name, 2},
     {"_Racmacs_ac_sr_set_name_full", (DL_FUNC) &_Racmacs_ac_sr_set_name_full, 2},
     {"_Racmacs_ac_sr_set_name_abbreviated", (DL_FUNC) &_Racmacs_ac_sr_set_name_abbreviated, 2},