+++ src/json_read_to_acmap.cpp
@@ -89,14 +89,13 @@ AcMap json_to_acmap(
     if(ag.HasMember("N")) map.antigens[i].set_name( ag["N"].GetString() );
     if(ag.HasMember("P")) map.antigens[i].set_passage( ag["P"].GetString() );
     if(ag.HasMember("c")) map.antigens[i].set_clade( parse<std::vector<std::string>>(ag["c"]) );
-    // set_group_values
-    // set_date
+    if(ag.HasMember("a")) map.antigens[i].set_annotations( parse<std::vector<std::string>>(ag["a"]) );
+    if(ag.HasMember("l")) map.antigens[i].set_labids( parse<std::vector<std::string>>(ag["l"]) );
+    if(ag.HasMember("A")) map.antigens[i].set_sequence( ag["A"].GetString() );
+    if(ag.HasMember("D")) map.antigens[i].set_date( ag["D"].GetString() );
     // set_reference
     // set_name_full
     // set_name_abbreviated
-    // set_id
-    // set_group
-    // set_sequence
 
   }
 
@@ -109,14 +108,13 @@ AcMap json_to_acmap(
     if(sr.HasMember("N")) map.sera[i].set_name( sr["N"].GetString() );
     if(sr.HasMember("P")) map.sera[i].set_passage( sr["P"].GetString() );
     if(sr.HasMember("c")) map.sera[i].set_clade( parse<std::vector<std::string>>(sr["c"]) );
-    // set_group_values
-    // set_date
+    if(sr.HasMember("a")) map.sera[i].set_annotations( parse<std::vector<std::string>>(sr["a"]) );
+    if(sr.HasMember("A")) map.sera[i].set_sequence( sr["A"].GetString() );
+    if(sr.HasMember("D")) map.sera[i].set_date( sr["D"].GetString() );
+    if(sr.HasMember("I")) map.sera[i].set_id( sr["I"].GetString() );
     // set_reference
     // set_name_full
     // set_name_abbreviated
-    // set_id
-    // set_group
-    // set_sequence
 
   }
 
@@ -269,7 +267,7 @@ AcMap json_to_acmap(
       for(SizeType i=0; i<xa.Size(); i++){
         const Value& xai = xa[i];
         if(xai.HasMember("g")) map.antigens[i].set_group( xai["g"].GetInt() );
-        if(xai.HasMember("q")) map.antigens[i].set_sequence( xai["q"].GetString() );
+        if(xai.HasMember("q")) map.antigens[i].set_sequence( xai["q"].GetString() ); // For backwards compatibility
         if(xai.HasMember("i")) map.antigens[i].set_id( xai["i"].GetString() );
         if(xai.HasMember("x")) map.antigens[i].set_extra( xai["x"].GetString() );
       }
@@ -281,8 +279,8 @@ AcMap json_to_acmap(
       for(SizeType i=0; i<xs.Size(); i++){
         const Value& xsi = xs[i];
         if(xsi.HasMember("g")) map.sera[i].set_group( xsi["g"].GetInt() );
-        if(xsi.HasMember("q")) map.sera[i].set_sequence( xsi["q"].GetString() );
-        if(xsi.HasMember("i")) map.sera[i].set_id( xsi["i"].GetString() );
+        if(xsi.HasMember("q")) map.sera[i].set_sequence( xsi["q"].GetString() ); // For backwards compatibility
+        if(xsi.HasMember("i")) map.sera[i].set_id( xsi["i"].GetString() ); // For backwards compatibility
         if(xsi.HasMember("x")) map.sera[i].set_extra( xsi["x"].GetString() );
       }
     }